test: move tests to 'right' path
