+++ test/node-test/diagnostics-channel/error.js
