+++ test/node-test/diagnostics-channel/get.js
