+++ test/node-test/diagnostics-channel/post-stream.js
